Fix adding events to existing charts
- Set an empty string default value for all required field of event-point
and event-range.
- Do not show unsupported date format warning if date field is empty.

The first change fixes a bug where it was not possible to add event / time span
annotations for existing charts.

If required fields do not have a default value, they are "undefined", and so a
newly created event does not validate against the event-point or event-range
schema.

Relevant code in Q-Editor:
https://github.com/nzzdev/Q-editor/blob/9ed126b2a9ca1fee214130c902c957f02f1ec79f/client/src/elements/schema-editor/schema-editor-array.js#L264-L273