Handle event annotations of intraday charts
Problem: Intraday charts do not have a linear time axis,
because the stock market is closed during the night.
Instead, there is a "jump" during the night, such that e.g.
Thursday 16:50 => 17:00 is the same step on the x axis as
Thursday 17:00 => Friday 8:00.

This situation was not properly handled for event annotations before and
is fixed here.

Event annotations are also added to the stock chart fixture for testing.